requirements.txt
discord.py
python-dotenv
yt-dlp
openai>=1.0.0
requests
aiohttp
lyricsgenius
asyncpraw
discord.py[voice]